Include requested URL on response object
The URL as requested is now saved on the response object under the 'url'
property.

In most cases, this will be the URL for the response resource, however,
in some cases it may not be. For instance, browsers do not expose the URL
of redirected requests. In that case, the URL before the redirect is
returned.

Issue: #123